requirements.txt
# Different, Together - QR 오더 서비스
# 필요한 패키지 목록

# Flask 및 확장 모듈
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.2

# 데이터베이스
SQLAlchemy==2.0.20
pymysql==1.1.0
cryptography==41.0.3

# 유틸리티
Werkzeug==2.3.7
Jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.0.0
Pillow==10.0.0
gunicorn==21.2.0

# API 및 서비스
google-cloud-translate==3.11.1

# QR 코드 생성
qrcode==7.4.2
Pillow==10.0.0